Port Crud Finished Update doesnt work

+++ backend/SpareHub/Repository/MySql/PortMySqlRepository.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using Persistence;
 using Persistence.MySql;
 using Repository.Interfaces;
-using Shared.DTOs.Vessel;
+using Shared.Exceptions;
 
 namespace Repository.MySql;
 
@@ -14,10 +14,10 @@ public class PortMySqlRepository(SpareHubDbContext dbContext, IMapper mapper) : 
     public async Task<Port> CreatePortAsync(Port port)
     {
         var portEntity = mapper.Map<PortEntity>(port);
-        dbContext.Ports.Add(portEntity);
+        await dbContext.Ports.AddAsync(portEntity);
         await dbContext.SaveChangesAsync();
-        port.Id = portEntity.Id.ToString();
-        return port;
+        
+        return mapper.Map<Port>(portEntity);
     }
 
     public async Task<List<Port>> GetPortsAsync()
@@ -38,15 +38,20 @@ public class PortMySqlRepository(SpareHubDbContext dbContext, IMapper mapper) : 
 
     public async Task UpdatePortAsync(string portId, Port port)
     {
+        
         var portEntity = mapper.Map<PortEntity>(port);
-        portEntity.Id = int.Parse(portId);
         dbContext.Ports.Update(portEntity);
         await dbContext.SaveChangesAsync();
     }
 
     public async Task DeletePortAsync(string portId)
     {
-        dbContext.Ports.Remove(new PortEntity { Id = int.Parse(portId) });
+        var portEntity = await dbContext.Ports.FirstOrDefaultAsync(p => p.Id == int.Parse(portId));
+        
+        if (portEntity == null)
+            throw new NotFoundException($"Port with id '{portId}' not found");
+        
+        dbContext.Ports.Remove(new PortEntity {Id = int.Parse(portId)});
         await dbContext.SaveChangesAsync();
     }
